Get rid of old schema (based on files/url's).
