+++ graphqlapi/dynamic_generation_getmeta.go
@@ -17,8 +17,8 @@ package graphqlapi
 import (
 	"fmt"
 
+	"github.com/creativesoftwarefdn/weaviate/database/schema"
 	"github.com/creativesoftwarefdn/weaviate/models"
-	"github.com/creativesoftwarefdn/weaviate/schema"
 	"github.com/graphql-go/graphql"
 )
