+++ graphqlapi/graphql_schema.go
@@ -17,8 +17,8 @@ package graphqlapi
 import (
 	"github.com/creativesoftwarefdn/weaviate/config"
 	dbconnector "github.com/creativesoftwarefdn/weaviate/database/connectors"
+	"github.com/creativesoftwarefdn/weaviate/database/schema"
 	"github.com/creativesoftwarefdn/weaviate/messages"
-	"github.com/creativesoftwarefdn/weaviate/schema"
 	"github.com/graphql-go/graphql"
 )
